Game controller Draft
This module takes in a 3-bit "done" signal from the RandomNum modules.
(This "done" signal is the sequence value taken from the LFSR that is
used in the RandomNum module).
This module also takes in an enable flag, and outputs to the LED
variable which is set to 3-bits as well (for 3 LEDs). If a time_out flag
is raised, no LEDs will light up.